requirements.txt
# 民航规章知识图谱 - 依赖包

# 图数据库
neo4j>=5.0.0

# HTTP 请求
requests>=2.28.0

# 环境变量管理
python-dotenv>=0.19.0

# 数据处理
pandas>=1.5.0

# 可选：异步处理
aiohttp>=3.8.0

# 可选：日志管理
loguru>=0.6.0
